revamped html, responsive css, implemented js
redid the html to make it fit the css better. added media queries and other small things to css. implemented the js files
i just want to make it clear that i (Chancelor) didn't create the css or js files; that was my groupmates Farhan and Danielle. i just cleaned it up a bit for the group.